\name{camelback2}
\Rdversion{1.1}
\alias{camelback2}

\title{ 2-dimensional test function "camelback2" }

\description{
Modification of the classical 2D camelback function. Inputs are defined in [0 1]x[0 1].
}

\usage{
camelback2(x)
}

\arguments{
  \item{x}{2-dimensional input (or matrix which lines or such inputs)}
  }
%\details{}
\value{Image of x by camelback2}
%\references{}
\author{ Victor Picheny, David Ginsbourger }
%\note{}

%\seealso{}

%\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory. 
\keyword{internal}
