\name{fun_test2d_ranjan}
\Rdversion{1.3}
\alias{fun_test2d_ranjan}

\title{ 2-dimensional test function "fun_test2d_ranjan" }

\description{
(x1,x2) -> x1 + x2 + x1*x2
}

\usage{
fun_test2d_ranjan(x)
}

\arguments{
  \item{x}{2-dimensional input (or matrix which lines or such inputs)}
}

\value{Image of x by fun_test2d_ranjan}
\references{Ranjan, P., Bingham, D., Michailidis, G.: Sequential experiment design for contour estimation from complex computer codes. Technometrics 50(4), 527-541 (2008)}
\author{ David Ginsbourger (Neuchatel University, Stochastic Hydrogeology Group) }

\keyword{internal}
