\name{fundet}
\Rdversion{1.1}
\alias{fundet}

\title{ 1-dimensional test function "fundet" }

\description{
sin(10*x)/(1+x)+2*cos(5*x)*x^3+0.841)/1.6
}

\usage{
fundet(x)
}

\arguments{
  \item{x}{1-dimensional input}
  }
%\details{}
\value{Image of x by fundet}
%\references{}
\author{ David Ginsbourger (Neuchatel University, Stochastic Hydrogeology Group) }
%\note{}

%\seealso{}

\examples{
curve(fundet)}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory. 
\keyword{internal}
