\name{lili2D}
\Rdversion{1.1}
\alias{lili2D}

\title{ 2-dimensional test function "lili2D" }

\description{
Inputs are defined in [0 1]x[0 1].
}

\usage{
lili2D(x)
}

\arguments{
  \item{x}{2-dimensional input (or matrix which lines or such inputs)}
  }
%\details{}
\value{Image of x by lili2D}
%\references{}
\author{ Victor Picheny, David Ginsbourger }
%\note{}

%\seealso{}

%\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory. 
\keyword{internal}
