\name{camelback2}
\Rdversion{1.3}
\alias{camelback2}

\title{ 2-dimensional test function "camelback2" }

\description{
Modification of the classical 2D camelback function. Inputs are defined in [0 1]x[0 1].
}

\usage{
camelback2(x)
}

\arguments{
  \item{x}{2-dimensional input (or matrix which lines or such inputs)}
}

\value{Image of x by camelback2}

\author{ Victor Picheny, David Ginsbourger }

\keyword{internal}
