\name{fundet}
\Rdversion{1.2}
\alias{fundet}

\title{ 1-dimensional test function "fundet" }

\description{
sin(10*x)/(1+x)+2*cos(5*x)*x^3+0.841)/1.6
}

\usage{
fundet(x)
}

\arguments{
  \item{x}{1-dimensional input}
  }

\value{Image of x by fundet}

\author{ David Ginsbourger (Neuchatel University, Stochastic Hydrogeology Group) }

\examples{
curve(fundet)}

\keyword{internal}

