\name{lili2D}
\Rdversion{1.2}
\alias{lili2D}

\title{ 2-dimensional test function "lili2D" }

\description{
Inputs are defined in [0 1]x[0 1].
}

\usage{
lili2D(x)
}

\arguments{
  \item{x}{2-dimensional input (or matrix which lines or such inputs)}
  }

\value{Image of x by lili2D}

\author{ Victor Picheny, David Ginsbourger }

\keyword{internal}
